consolidates all essential images to src/img and adds body-parts background imgs to src/imgs
